+++ packages/rpc-provider/src/lru.ts
@@ -20,7 +20,8 @@ class LRUNode {
 
 // https://en.wikipedia.org/wiki/Cache_replacement_policies#LRU
 export class LRUCache {
-  readonly capacity: number;
+  public readonly capacity: number;
+
   readonly #data: Map<string, unknown> = new Map();
   readonly #refs: Map<string, LRUNode> = new Map();