+++ packages/types-codec/src/base/Enum.ts
@@ -183,22 +183,22 @@ function decodeEnum (registry: Registry, def: TypesDef, value?: unknown, index?:
  * an extension to enum where the value type is determined by the actual index.
  */
 export class Enum implements IEnum {
+  /** @deprecated This is not populated anymore. Use $createdAtHash instead. */
+  public createdAtHash?: never;
+  /** @deprecated This is not populated anymore. Use $initialU8aLength instead. */
+  public initialU8aLength?: never;
+
   public readonly registry: Registry;
 
-  public createdAtHash?: IU8a;
+  public $createdAtHash?: IU8a;
+  public $initialU8aLength?: number;
+  public $isStorageFallback?: boolean;
 
   readonly #def: TypesDef;
-
   readonly #entryIndex: number;
-
-  readonly initialU8aLength?: number;
-
   readonly #indexes: number[];
-
   readonly #isBasic: boolean;
-
   readonly #isIndexed: boolean;
-
   readonly #raw: Codec;
 
   constructor (registry: Registry, Types: Record<string, string | CodecClass> | Record<string, number> | string[], value?: unknown, index?: number, { definition, setDefinition = noopSetDefinition }: Options = {}) {
@@ -217,8 +217,8 @@ export class Enum implements IEnum {
     this.#entryIndex = this.#indexes.indexOf(decoded.index);
     this.#raw = decoded.value;
 
-    if (this.#raw.initialU8aLength) {
-      this.initialU8aLength = 1 + this.#raw.initialU8aLength;
+    if (this.#raw.$initialU8aLength) {
+      this.$initialU8aLength = 1 + this.#raw.$initialU8aLength;
     }
   }