+++ packages/types-codec/src/base/Tuple.ts
@@ -55,8 +55,6 @@ function decodeTuple (registry: Registry, result: Codec[], value: Exclude<AnyTup
  * own type. It extends the base JS `Array` object.
  */
 export class Tuple extends AbstractArray<Codec> implements ITuple<Codec[]> {
-  readonly initialU8aLength?: number;
-
   #Types: Definition;
 
   constructor (registry: Registry, Types: TupleTypes | TupleType, value?: AnyTupleValue, { definition, setDefinition = noopSetDefinition }: Options = {}) {
@@ -70,7 +68,7 @@ export class Tuple extends AbstractArray<Codec> implements ITuple<Codec[]> {
 
     super(registry, Classes[0].length);
 
-    this.initialU8aLength = (
+    this.$initialU8aLength = (
       isU8a(value)
         ? decodeU8a(registry, this, value, Classes)
         : decodeTuple(registry, this, value, Classes)